% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborNet.R
\name{neighborNet}
\alias{neighborNet}
\title{Neighbor Net: An approach to infer putative disease-specific mechanisms using neighboring gene networks.}
\usage{
neighborNet(de, ref, listofgenes, threshold = 0.1, minsize = 2)
}
\arguments{
\item{de}{a vector including the differentially expressed genes; \code{de} must use the same id's as \code{ref} and the genes in the \code{listofgenes}}

\item{ref}{the reference vector for all genes in the analysis}

\item{listofgenes}{a list representing the neighbor networks associated to each gene; the name of the list must be the same as genes in the \code{de}}

\item{threshold}{a threshold of choosing significant neighbor networks (default is 0.1)}

\item{minsize}{minimum size of the neighbor networks that should be considered in the analysis (default is 2)}
}
\value{
An object of class \code{\link{graphNEL}}.
}
\description{
Neighbor Net: An approach to infer putative disease-specific mechanisms using neighboring gene networks.
}
\details{
See details in the cited articles.
}
\examples{

# load multiple colorectal cancer study (public data available in GEO
# ID: GSE4183, GSE9348, GSE21510, GSE32323, GSEl8671)
# These files contains the tables, produced by the limma package with
# added gene information.
# The table contains the expression fold change and signficance of each
# probe set comparing colorectal cancer disease and normal.
load(system.file("extdata/dataColorectal4183.RData", package = "NeighborNet"))
load(system.file("extdata/dataColorectal9348.RData", package = "NeighborNet"))
load(system.file("extdata/dataColorectal21510.RData", package = "NeighborNet"))
load(system.file("extdata/dataColorectal32323.RData", package = "NeighborNet"))
load(system.file("extdata/dataColorectal8671.RData", package = "NeighborNet"))
head(dataColorectal4183)

load(system.file("extdata/listofgenes.RData", package = "NeighborNet"))
head(listofgenes)

# select differentially expressed genes for each data set at p-value below 1\%
# and absolute value for more than 1.5 and save their entrez ID in a vector de1 to de5
pvThreshold <- 0.01
foldThreshold <- 1.5
de1 <- dataColorectal4183$EntrezID [
  dataColorectal4183$adj.P.Val < pvThreshold &
  abs(dataColorectal4183$logFC) > foldThreshold]
de2 <- dataColorectal9348$EntrezID [
  dataColorectal9348$adj.P.Val < pvThreshold &
  abs(dataColorectal9348$logFC) > foldThreshold]
de3 <- dataColorectal21510$EntrezID [
  dataColorectal21510$adj.P.Val < pvThreshold &
  abs(dataColorectal21510$logFC) > foldThreshold]
de4 <- dataColorectal32323$EntrezID [
  dataColorectal32323$adj.P.Val < pvThreshold &
  abs(dataColorectal32323$logFC) > foldThreshold]
de5 <- dataColorectal8671$EntrezID [
  dataColorectal8671$adj.P.Val < pvThreshold &
  abs(dataColorectal8671$logFC) > foldThreshold]
all <- unique( c(dataColorectal4183$EntrezID, dataColorectal9348$EntrezID,
  dataColorectal21510$EntrezID, dataColorectal32323$EntrezID,
  dataColorectal8671$EntrezID))
de <- unique( c(de1,de2,de3,de4,de5))

sig_net <- neighborNet (de, all, listofgenes)

}
\references{
Sahar Ansari, Michele Donato, Nafiseh Saberian, Sorin Draghici; An approach to infer putative disease-specific mechanisms using neighboring gene networks, Bioinformatics, Volume 33, Issue 13, 1 July 2017, Pages 1987–1994
}
\author{
Sahar Ansari and Sorin Draghici
}
